requirements.txt
yfinance>=0.2.18
mplfinance>=0.12.9b0
pandas>=1.5.3
numpy>=1.24.3
matplotlib>=3.7.1
pyyaml>=6.0
python-telegram-bot>=13.7
apscheduler>=3.9.1 